+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Custom auth and header parameters for `FhirServer` http session.
 - `FhirServer` constructor now accepts two additional optional parameters, `auth` and `headers` that will be used for the
   instance's `requests` session.
 
+### Added
+ - [x] `server.get(reference)` get a single resource from the server, based on relative path/reference.
+ - [x] `server.get_many(references)` get multiple resources from the server, based on relative path/reference.
 
 ## [0.7.0] - 2022-01-31
 Update a list of resources on the server. CSV/Pandas serialization of resources and query responses.